refactor(routers): split analytics and admin routers into directory modules
Follows the established crawler/files router pattern:
- analytics/ → index + helpers + 5 procedure files (usage, sessions,
  insights, conversations, export-conversations)
- admin/ → index + helpers (requireExistingUser, sendOptionalEmail
  replacing the 5x user-fetch and 4x email try/catch copies) +
  6 procedure files

No logic changes; every error message, email flow and return shape
preserved verbatim. Largest file now 281 lines.
